engine/requirements.txt
playwright==1.42.0
google-generativeai==0.4.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
